🚚Distribute section class defs across files
Spread out the classes more across files in order to reduce file lengths
 which were becoming unwieldy.
